Pong
Finished another youtube tutorial, I changed a few things on how I implement the same concepts, along with adding comments to clarify what each piece of code does. This will be my final Python project for a while, I think I understand the concepts, and now it is time to move on to JavaScript, I will post updates on my projects as time goes on!